Support more than one chat connection at a time (#194)
* Support more than one chat connection at a time

* Move some things around

* Don't hardcode port
